requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
benepar==0.2.0
bleach==5.0.0
blinker==1.4
blis==0.7.7
boto3==1.24.66
botocore==1.27.66
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
docutils==0.19
emoji==1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.14.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keyring==23.6.0
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pickleshare==0.7.5
Pillow==9.1.1
pkginfo==1.8.3
plotly==5.8.2
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytorch-transformers==1.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.1.0
readme-renderer==35.0
regex==2022.6.2
requests==2.28.0
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
s3transfer==0.6.0
sacremoses==0.0.53
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.96
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spacy-streamlit==1.0.4
srsly==2.4.3
st-dep-parsing-component==0.0.3
st-parsetree==0.0.2
st-tokenization-vis-component==0.0.8
stack-data==0.2.0
stanza==1.4.0
streamlit==1.12.2
tenacity==8.0.1
terminado==0.15.0
thinc==8.0.17
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
toolz==0.11.2
torch==1.11.0
torch-struct==0.5
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
transformers==4.19.4
twine==4.0.1
typer==0.4.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
wasabi==0.9.1
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
wget==3.2
widgetsnbextension==3.6.0
yarg==0.1.9
zipp==3.8.0
